Add files via upload
"gui.py" introduces the 1st version of a gui for ChefAI. Allowing the user to add preferences and foods that they currently have, to the TKinter gui. TKinter will create a JSON file, containing the URL, image and nutritional info for the food item. Want to expand this out to provide a list, and have an intermediary step to show a collated list, of "X" number of 'hits' generated.